More response definition checking
Now checks whether "default" response is defined. If security is defined (globally or locally) for the operation, checks that 401 and 403 response codes are defined.

checkSecurityField() now takes the object to be checked as an argument. responseCheck() takes as an argument a boolean that describes whether global security is defined (true) or not (false).